Introduction and recipe link list in the index, and boilerplate and tittles in the recipes.
In the index page I added a brief introduction regarding on what the page is about, and added a ordered list containing all the recipes with links to those pages ona blank tab.

In the recipe folder, I added the bolerplate to all the recipes and their respective titles on the tab name.