Changed index and added pizza recipe.html
In the index page, I divided the recepies into sweet and savoury adding the pizza page and link to the index. I also fixed the mistake I made, where the recipe for the pasta alfredo was anchored to pancakes.

To the pizza file, I only added the boilerplate and correct title.